requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.7
django-redis==5.4.0
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
kombu==5.3.7
promise==2.3
prompt_toolkit==3.0.47
psycopg2==2.9.9
python-dateutil==2.9.0.post0
redis==5.0.7
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
